MorriMoreBOT/requirements.txt
python-telegram-bot==20.0
requests
colorama
concurrent.futures; python_version < "3.9"